Re-enable prospector hook and fix reported issues
