+++ .idea/dictionaries/idlecode.xml
@@ -2,6 +2,8 @@
   <dictionary name="idlecode">
     <words>
       <w>callables</w>
+      <w>isort</w>
+      <w>pylint</w>
       <w>rabarbar</w>
       <w>subnode</w>
       <w>subnodes</w>
