+++ .pre-commit-config.yaml
@@ -8,16 +8,6 @@ repos:
       - id: debug-statements
       - id: end-of-file-fixer
       - id: trailing-whitespace
-  - repo: https://github.com/pre-commit/mirrors-mypy
-    rev: master
-    hooks:
-      - id: mypy
-        files: 'contextshell|tests'
-#  - repo: https://github.com/guykisel/prospector-mirror
-#    rev: master
-#    hooks:
-#      - id: prospector
-#        exclude: "tests"
   - repo: https://github.com/asottile/pyupgrade
     rev: master
     hooks:
@@ -28,6 +18,9 @@ repos:
     rev: master
     hooks:
       - id: isort
+        args:
+          - --multi-line=3
+          - --keep-direct-and-as
   - repo: https://github.com/ambv/black
     rev: master
     hooks:
@@ -36,3 +29,13 @@ repos:
         args:
           - -l 100
           - --py36
+  - repo: https://github.com/guykisel/prospector-mirror
+    rev: master
+    hooks:
+      - id: prospector
+        exclude: "tests"
+  - repo: https://github.com/pre-commit/mirrors-mypy
+    rev: master
+    hooks:
+      - id: mypy
+        files: 'contextshell|tests'
