+++ .pylintrc
@@ -525,7 +525,7 @@ max-locals=10
 max-parents=5
 
 # Maximum number of public methods for a class (see R0904).
-max-public-methods=20
+max-public-methods=30
 
 # Maximum number of return / yield for function / method body.
 max-returns=6
