+++ tests/functional/CrudTests.py
@@ -1,6 +1,7 @@
+import unittest
+
 from tests.functional.ShellTestsBase import NodeTreeTestsBase
 from tests.functional.TestExecutor import script_test
-import unittest
 
 
 class CrudTestsBase(NodeTreeTestsBase):
