+++ tests/unit/ActionExecutorTests.py
@@ -2,14 +2,14 @@ import unittest
 from collections import OrderedDict
 from typing import Optional
 
-from contextshell.action import Action, pack_argument_tree
-from contextshell.path import NodePath, NodePath as np
-from contextshell.action import ActionExecutor
+from contextshell.action import Action, ActionExecutor, pack_argument_tree
 from tests.unit.Fakes import FakeAction
+
+from contextshell.path import NodePath, NodePath as np  # isort:skip
 
 
 class TestableActionExecutor(ActionExecutor):
-    found_action: Action = None
+    found_action: Optional[Action] = None
 
     def find_action(self, target: NodePath, action: NodePath) -> Optional[Action]:
         return self.found_action
