+++ tests/unit/NodePathTests.py
@@ -1,4 +1,5 @@
 import unittest
+
 from contextshell.path import *
